first commit
There are three untracked files. One is script.properties - that contains
sensitive information about connection to the server. This file is
documented in the README.md - so that one could just copy-paste it
into a blank file called script.properties and then edit according to
one's local limesurvey parameters.

The other file is README.pdf created by pandoc from README.md

The third file is patientList-demo.pdf - the file with URL QR-codes
specific to my installation of limesurvey. I have briefly explained how I
produce this file, but this is beyond scope of this project. People
can make QR codes or distribute links to their survey in many ways.

Changes to be committed:
	new file:   .gitignore
	new file:   README.md
	new file:   fbHWU-01-DEMO.pdf
	new file:   fbHWU-01-DEMO.rmd
	new file:   resetDemoData.R
	new file:   survey_structure/limesurvey_survey_DEMO.lss
	new file:   survey_structure/vvexport_467000.csv